[Refactor] move `FeedImagesMapper` class to its own file
